refactor: extract handler for usage.
